+++ blocks/init/custom/divider/docs/story.js
@@ -1,0 +1,29 @@
+import { Gutenberg, blockDetails, hasWrapperDecorator } from 'EighshiftBlocksStorybookHelpers';
+import React from 'react'; // eslint-disable-line no-unused-vars
+import manifest from '../manifest.json';
+import readme from './readme.md';
+
+export default {
+  title: 'Blocks|Divider',
+  ...hasWrapperDecorator(manifest),
+  parameters: {
+    notes: readme,
+  },
+};
+
+export const block = () => (
+  <Gutenberg props={
+    {
+      blocks: [
+        {
+          attributes: {
+            ...blockDetails(manifest.blockName),
+            color: 'primary',
+          },
+          innerBlocks: [],
+          name: `eightshift-boilerplate/${manifest.blockName}`,
+        },
+      ],
+    }
+  } />
+);